refactor Position into a shared component
