+++ visualizations/leaderboard/styles.scss
@@ -2,7 +2,7 @@
 @import "./styles/LeaderboardItem.scss";
 @import "./styles/ProgressBar.scss";
 @import "./styles/ChangeIndicator.scss";
-@import "./styles/Position.scss";
+@import "../../shared_components/Position/Position.scss";
 
 // ErrorState.tsx and EmptyState.tsx
 .EmptyState,
