+++ visualizations/podium/styles.scss
@@ -1,4 +1,4 @@
-@import "../leaderboard/styles/Position.scss";
+@import "../../shared_components/Position/Position.scss";
 
 .EmptyState,
 .ErrorState {
